+++ src/index.ts
@@ -1,11 +1,6 @@
-export { default as computeTooltipCoords } from './computeTooltipCoords';
-export { default as mergeProps } from './mergeProps';
+export { default as Portal } from './Portal';
 export { default as PortalNodeContext } from './PortalNodeContext';
 export { default as Tooltip } from './Tooltip';
-export { default as TooltipParent } from './TooltipParent';
-export { default as useClickTooltip } from './useClickTooltip';
-export { default as useFocusTooltip } from './useFocusTooltip';
-export { default as useHoverTooltip } from './useHoverTooltip';
 export { default as useOutsideClick } from './useOutsideClick';
 export { default as useTooltip } from './useTooltip';
 export { default as ZIndexContext } from './ZIndexContext';